Sửa lỗi từ hai lượt review độc lập
LỚP NATIVE — hai lỗi bảo đảm chết lần chạy đầu:

SetDllImportResolver chỉ có tác dụng cho ĐÚNG MỘT assembly. MpvRenderNative nằm
ở Nocturne.Render còn resolver đăng ký cho Nocturne.Engine, nên mọi lời gọi
render API sẽ đi dò "mpv.dll" — file không tồn tại trong bất kỳ bản libmpv nào.
Thêm MpvRuntime.RegisterInteropAssembly và gọi trước lời gọi đầu tiên.

Cách đưa D3D11 device cho ANGLE hoàn toàn sai. EGL_D3D11_DEVICE_ANGLE là token
tạo device, không phải display attribute; nhét nó vào attrib list thì tốt nhất
là ANGLE lặng lẽ tự tạo device thứ hai — phá đúng bất biến "một device" mà cả
pipeline dựa vào, và làm eglCreatePbufferFromClientBuffer fail vì texture thuộc
device khác. Đường đúng: eglCreateDeviceANGLE rồi eglGetPlatformDisplayEXT với
EGL_PLATFORM_DEVICE_EXT. Kèm theo: eglGetPlatformDisplayEXT nhận EGLint* 32-bit
chứ không phải EGLAttrib*, chú thích cũ ghi ngược.

Hai lỗi lúc tắt: mpv_render_context_free đòi GL context phải current (nó huỷ
shader/FBO/texture cache), mà render thread đã clear trước khi thoát. Và khi
join hết 2 giây thì code vẫn dọn sạch tài nguyên dưới chân một thread còn sống —
đúng thứ mà chính chú thích ở đó tuyên bố sẽ không làm. Giờ thà leak có chủ đích.

Thêm: mpv_wakeup gọi SAU khi xin write lock nên không bao giờ tới kịp, mọi lần
đóng cửa sổ đơ tới 1 giây; playlist_insert_id là int64_t chứ không phải int nên
field sau nó đọc nửa cao của giá trị; bỏ fallback mpv-1.dll vì struct
mpv_opengl_init_params của nó có 3 field; chặn tự-join khi Dispose gọi từ chính
pump thread; bỏ EGL_WIDTH/HEIGHT khỏi pbuffer từ D3D texture.

LỚP APP — cụm seek bar hỏng hoàn toàn:

RangeBase.ValueChanged bắn cho cả thay đổi programmatic, và FocusState không nói
gì về NGUỒN của thay đổi. Chỉ cần user click seek bar một lần là mỗi tick engine
cập nhật Progress sẽ kích một lệnh seek — video giật liên tục, không xem nổi. Bỏ
hẳn đường commit qua ValueChanged; phím mũi tên xử lý riêng.

Tệ hơn: Slider đánh dấu Handled cho PointerPressed/Released để capture con trỏ,
nên handler gắn trong XAML KHÔNG BAO GIỜ chạy — cơ chế chống thumb-snap-back
trông như đã nối mà thực ra im lặng. Phải dùng AddHandler với handledEventsToo.
Thêm PointerCaptureLost để một cú kéo bị alt-tab không khoá cứng cờ scrub.

Lỗi render bị snapshot thường đè mất trong vài mili giây — trên máy không có
ANGLE, tức là mọi máy hiện giờ, nó đọc thành "app mở lên rồi không làm gì".
Giờ chốt lại, không cho đè.

KeyDown chỉ bắn khi có element giữ focus; lúc mới mở transport bar đang
Collapsed nên không phím tắt nào sống, và sau khi bấm một nút thì Space bấm lại
đúng nút đó. Chuyển sang KeyboardAccelerator. Thêm Ctrl+O — màn hình rỗng vẫn
đang bảo user bấm nó suốt mà chưa hề có.

Kèm: playlist bị mutate từ thread pump trước khi hop sang UI thread; volume
slider dính đúng vòng lặp phản hồi như seek bar; async void OnDrop không có
catch nên một payload drag ảo giết cả process; toàn màn hình còn nguyên dải
title bar 40px; queue lệch khi mở file có đuôi lạ; AppWindowTitleBar không guard
IsCustomizationSupported.

Thêm 3 test: duration âm từ container hỏng giờ được gập về "không rõ độ dài".

+++ src/Nocturne.Engine/Interop/MpvStructs.cs
@@ -49,7 +49,10 @@ internal struct MpvEventEndFile
     public MpvEndFileReason Reason;
     public int Error;
     public long PlaylistEntryId;
-    public int PlaylistInsertId;
+
+    // int64_t in client.h, not int. Declaring it as int shifts every field after
+    // it, so PlaylistInsertNumEntries would read the high half of this value.
+    public long PlaylistInsertId;
     public int PlaylistInsertNumEntries;
 }
